fix join_equivalence_properties
